Moved rm_rf_tmp to test-utils.
